FIX: FP16 softmax overflow causing NaN cascade
ROOT CAUSE #2: Softmax computing in FP16

PROBLEM:
Spectral attention scores can reach ±300 raw, ±26 after scaling.
In softmax: exp(26) ≈ 5×10^11, but FP16 max is 65504.
exp(11) is near FP16 max, so exp(26) OVERFLOWS TO INF!

Inf → NaN → Propagates through model → SVD fails on NaN data

FILES FIXED:
1. kernels/spectral_attention.py (line 802)
2. spectral_attention.py (line 288, 381)

BEFORE:
  attn_weights = F.softmax(scores, dim=-1)  # FP16 overflow!

AFTER:
  attn_weights = F.softmax(scores, dim=-1, dtype=torch.float32).to(scores.dtype)

NOTE: integration.py already had this pattern correct, which is why
standard attention worked. Only spectral attention paths were broken.

BUGS FIXED THIS SESSION:
1. Hot cache RoPE: K * cos → _apply_rope(K, cos, sin)
2. rSVD FP32: Added Half→Float conversion for QR decomposition
3. Softmax FP32: Prevent exp() overflow in FP16

+++ unsloth_spectral/kernels/spectral_attention.py
@@ -799,7 +799,10 @@ def spectral_attention_decode(
     if attention_mask is not None:
         scores = scores.masked_fill(attention_mask == 0, float('-inf'))
     
-    attn_weights = F.softmax(scores, dim=-1)  # [B, H_q, 1, T_total]
+    # CRITICAL: Compute softmax in FP32 to prevent exp() overflow!
+    # Scores can reach ±26 after scaling, and exp(26) ≈ 5×10^11 overflows FP16 (max 65504)
+    # This causes Inf → NaN cascade that corrupts the entire model
+    attn_weights = F.softmax(scores, dim=-1, dtype=torch.float32).to(scores.dtype)  # [B, H_q, 1, T_total]
     
     if config.enable_logging:
         logger.debug(f"[PHASE2] Softmax applied: attn_sum={attn_weights.sum(dim=-1).mean():.4f}")
